Use index instead of broker
